Fixing a couple of warnings...
git-svn-id: https://svn.atnf.csiro.au/askapsdp/trunk@1545 1db04d4c-0ff0-4329-ae0b-98c546476c6a

+++ Code/Components/Analysis/analysis/trunk/sourcefitting/Fitter.cc
@@ -73,7 +73,7 @@ namespace askap
 	this->itsNumGauss = parset.getInt32("numGauss", defaultNumFittedGauss);
 	this->itsBoxPadSize = parset.getInt32("boxPadSize", detectionBorder);
 	this->itsChisqCutoff = parset.getFloat("chisqCutoff", defaultChisqCutoff);
-	this->itsNoiseBoxSize = parset.getFloat("noiseBoxSize", defaultNoiseBoxSize);
+	this->itsNoiseBoxSize = parset.getInt32("noiseBoxSize", defaultNoiseBoxSize);
 	
 	this->itsFitter.setDimensions(2);
